fix(qe): don't crash on relationLoadStrategy in compacted batches
