+++ query-compiler/query-engine-tests-todo/pg-cockroachdb/fail/query
@@ -1,4 +1,6 @@
 new::regressions::prisma_7434::not_in_chunking_cockroachdb::not_in_batch_filter
+new::relation_load_strategy::batch::relation_load_strategy_batch::compacted_join_lateral
+new::relation_load_strategy::batch::relation_load_strategy_batch::compacted_query_lateral
 queries::aggregation::group_by::aggregation_group_by::group_by_ordering_sum_aggregation
 queries::chunking::chunking::order_by_aggregation_should_fail
 queries::chunking::chunking::order_by_relevance_should_fail
