+++ query-compiler/query-engine-tests-todo/planetscale/fail/query
@@ -1,5 +1,7 @@
 new::interactive_tx::interactive_tx::batch_queries_failure
 new::regressions::prisma_7434::not_in_chunking::not_in_batch_filter
+new::relation_load_strategy::batch::relation_load_strategy_batch::compacted_join_subquery
+new::relation_load_strategy::batch::relation_load_strategy_batch::compacted_query_subquery
 queries::aggregation::group_by::aggregation_group_by::group_by_ordering_sum_aggregation
 queries::chunking::chunking::order_by_aggregation_should_fail
 queries::filters::self_relation_regression::sr_regression::all_categories
